render dashboards page
