+++ app/javascript/app/layouts/root/styles.scss
@@ -1,6 +1,7 @@
 .l-root {
   .page {
     position: relative;
+    z-index: 0;
 
     .page-loader {
       height: calc(100vh - 56px);
