+++ app/javascript/app/pages/dashboards/header/header-styles.scss
@@ -1,6 +1,6 @@
 @import '~styles/settings.scss';
 
-.c-header {
+.c-header-menu {
   position: relative;
   background-color: $slate-light;
   padding: rem(50px) 0 rem(30px);
