+++ app/javascript/app/pages/dashboards/reducers.js
