+++ app/javascript/app/pages/dashboards/styles.scss
