+++ app/javascript/app/reducers.js
@@ -13,6 +13,9 @@ import * as ModalMetaComponent from 'components/modals/meta';
 import * as WidgetsComponent from 'components/widgets';
 import * as PopupComponent from 'components/map-v2/components/popup';
 
+import * as HeaderComponent from 'pages/dashboards/header';
+import * as MapComponent from 'components/map';
+
 // Providers
 import * as CountryDataProviderComponent from 'providers/country-data-provider';
 import * as GeostoreProviderComponent from 'providers/geostore-provider';
@@ -20,6 +23,10 @@ import * as WhitelistsProviderComponent from 'providers/whitelists-provider';
 import * as DatasetsProviderComponent from 'providers/datasets-provider';
 import * as MyGFWProviderComponent from 'providers/mygfw-provider';
 import * as PTWProviderComponent from 'providers/ptw-provider';
+import * as LayerSpecProviderComponent from 'providers/layerspec-provider';
+
+// pages
+import * as DashboardsPageComponent from 'pages/dashboards';
 
 // Component Reducers
 const componentsReducers = {
@@ -28,7 +35,9 @@ const componentsReducers = {
   dataAnalysis: handleActions(DataAnalysisMenuComponent),
   recentImagery: handleActions(RecentImageryComponent),
   widgets: handleActions(WidgetsComponent),
-  popup: handleActions(PopupComponent)
+  popup: handleActions(PopupComponent),
+  header: handleActions(HeaderComponent),
+  map: handleActions(MapComponent)
 };
 
 // Provider Reducers
@@ -38,11 +47,18 @@ const providersReducers = {
   whitelists: handleActions(WhitelistsProviderComponent),
   datasets: handleActions(DatasetsProviderComponent),
   myGfw: handleActions(MyGFWProviderComponent),
-  ptw: handleActions(PTWProviderComponent)
+  ptw: handleActions(PTWProviderComponent),
+  layerSpec: handleActions(LayerSpecProviderComponent)
+};
+
+// Page Reducers
+const pageReducers = {
+  dashboards: handleActions(DashboardsPageComponent)
 };
 
 export default combineReducers({
   ...providersReducers,
   ...componentsReducers,
+  ...pageReducers,
   location: router.reducer
 });
