+++ app/views/react/index.html.erb
@@ -1,5 +1,6 @@
 <%= content_for :head do %>
   <%= render 'shared/leaflet' %>
+  <%= render 'shared/google_carto' %>
 <% end %>
 
 <div id="app-root"></div>
